aider: Changed the way products and quantities are accessed in the orders.html template to display the product names and total quantity for each order.
# Context:
USER: The table that displays the generated orders isn't showing anything in the products column nor the quanity column
ASSISTANT: To help you with this, I need to see the code for the table that displays the generated orders. Could you please provide that?
USER: I added these *read-write* files: accounting/templates/order_detail.html
ASSISTANT: I see that the order detail template is provided, but I also need to see the template that displays the list of orders (the table you mentioned). Could you please provide that as well?
USER: I added these *read-write* files: accounting/templates/orders.html
ASSISTANT: Thank you for providing the necessary file. It seems like the issue is with the way the products and their quantities are being accessed in the template.

In the `orders.html` file, the products and their quantities are being accessed through `order.products.all`, but it seems like the `Order` object doesn't have a `products` attribute. Instead, it has a `product_names` attribute.

Similarly, the quantity is being accessed through `product.quantity`, but it seems like the `Order` object has a `quantity` attribute that represents the total quantity of all products in the order.

Here are the changes that need to be made:

```html
accounting/templates/orders.html
<<<<<<< ORIGINAL
            <td>
                {% for product in order.products.all %}
                    {{ product.name }}<br>
                {% endfor %}
            </td>
            <td>
                {% for product in order.products.all %}
                    {{ product.quantity }}<br>
                {% endfor %}
            </td>
=======
            <td>{{ order.product_names }}</td>
            <td>{{ order.quantity }}</td>
>>>>>>> UPDATED
```

This change will display the product names and the total quantity for each order. If you want to display the quantity for each product separately, you would need to modify the `OrderGenerator` to store the quantity for each product separately.