aider: Reintroduced the `order_quantity` variable and used it to determine the size of the `product_names` array in the `generate_order` method of the `OrderGenerator` class.
# Context:
USER: NameError at /accounting/generate_orders/
name 'order_quantity' is not defined
Request Method:	POST
Request URL:	http://127.0.0.1:8000/accounting/generate_orders/
Django Version:	3.2.9
Exception Type:	NameError
Exception Value:
name 'order_quantity' is not defined
Exception Location:	/Users/drakestapleton/PycharmProjects/Django-CRM/accounting/order_generator.py, line 33, in generate_order
Python Executable:	/opt/homebrew/Caskroom/miniconda/base/envs/django/bin/python
Python Version:	3.11.3
Python Path:
['/Users/drakestapleton/PycharmProjects/Django-CRM',
 '/opt/homebrew/Caskroom/miniconda/base/envs/django/lib/python311.zip',
 '/opt/homebrew/Caskroom/miniconda/base/envs/django/lib/python3.11',
 '/opt/homebrew/Caskroom/miniconda/base/envs/django/lib/python3.11/lib-dynload',
 '/Users/drakestapleton/.local/lib/python3.11/site-packages',
 '/opt/homebrew/Caskroom/miniconda/base/envs/django/lib/python3.11/site-packages']
Server time:	Thu, 03 Aug 2023 16:58:54 +0000
ASSISTANT: The error message indicates that the variable `order_quantity` is not defined in the `generate_order` method of the `OrderGenerator` class in the `accounting/order_generator.py` file.

This is because the `order_quantity` variable was removed in the previous changes, but it is still being used to determine the size of the `product_names` array.

To fix this, we need to reintroduce the `order_quantity` variable and use it to determine the size of the `product_names` array.

Here are the changes that need to be made:

```python
accounting/order_generator.py
<<<<<<< ORIGINAL
    def generate_order(self):
        ...
        product_names = np.random.choice(self.PRODUCTS, size=order_quantity, replace=False,
                                         p=self._get_product_probabilities())
        quantities = [int(self._get_order_quantity()) for _ in range(len(product_names))]
        product_names = ', '.join(f'{name} ({quantity})' for name, quantity in zip(product_names, quantities))
        ...
        total_price = sum([self.PRICES[self.PRODUCTS.index(product_name.split(' ')[0])] * quantity for product_name, quantity in zip(product_names.split(', '), quantities)])
        return {'customer_name': customer_name, 'product_names': product_names, 'order_date': order_date_str,
                'total_price': total_price}
=======
    def generate_order(self):
        ...
        order_quantity = int(self._get_order_quantity())
        product_names = np.random.choice(self.PRODUCTS, size=order_quantity, replace=False,
                                         p=self._get_product_probabilities())
        quantities = [int(self._get_order_quantity()) for _ in range(len(product_names))]
        product_names = ', '.join(f'{name} ({quantity})' for name, quantity in zip(product_names, quantities))
        ...
        total_price = sum([self.PRICES[self.PRODUCTS.index(product_name.split(' ')[0])] * quantity for product_name, quantity in zip(product_names.split(', '), quantities)])
        return {'customer_name': customer_name, 'product_names': product_names, 'order_date': order_date_str,
                'total_price': total_price}
>>>>>>> UPDATED
```

This change reintroduces the `order_quantity` variable and uses it to determine the size of the `product_names` array. The rest of the method remains the same.