aider: Fix split issue in generate_order method of OrderGenerator class.
# Context:
USER: ValueError at /accounting/generate_orders/
'Body' is not in list
Request Method:	POST
Request URL:	http://127.0.0.1:8000/accounting/generate_orders/
Django Version:	3.2.9
Exception Type:	ValueError
Exception Value:
'Body' is not in list
ASSISTANT: The error message indicates that the string 'Body' is not found in the list of products. This error is likely occurring in the `generate_order` method of the `OrderGenerator` class in the `accounting/order_generator.py` file, specifically in the line where the total price is calculated.

The issue is that the product names are being split on ', ', which results in product names like 'Body (1)' being split into 'Body' and '(1)'. When 'Body' is used to find the index of the product in the `PRODUCTS` list, it is not found because the actual product name is 'Body Wash'.

To fix this, we need to split the product names on ' (' instead of ', '.

Here are the changes that need to be made:

```python
accounting/order_generator.py
<<<<<<< ORIGINAL
    def generate_order(self):
        ...
        total_price = sum([self.PRICES[self.PRODUCTS.index(product_name.split(' ')[0])] * quantity for product_name, quantity in zip(product_names.split(', '), quantities)])
        return {'customer_name': customer_name, 'product_names': product_names, 'order_date': order_date_str,
                'total_price': total_price}
=======
    def generate_order(self):
        ...
        total_price = sum([self.PRICES[self.PRODUCTS.index(product_name.split(' (')[0])] * quantity for product_name, quantity in zip(product_names.split(', '), quantities)])
        return {'customer_name': customer_name, 'product_names': product_names, 'order_date': order_date_str,
                'total_price': total_price}
>>>>>>> UPDATED
```

This change splits the product names on ' (' instead of ', ', which ensures that the full product name is used to find the index of the product in the `PRODUCTS` list.